MoviesPresenter creation of data package with nested packages and moving relevant classes
